Make EventWaiter allow async predicates
Allow, but *not* require. Why don't more people do that?
This is a bit of an ugly hack, involving Task.FromResult(). This shouldn't be any *more* expensive
than making a function async that doesn't need to be though. Also, LINQ does not support async lambdas,
so change to using hand-rolled loops. Quite inconvenient.